Cambios en la configuración de jest, depuración de funciones no usadas.
