+++ package.json
@@ -17,7 +17,7 @@
         "url": "git+https://github.com/Tania1295/BOG005-md-links.git"
     },
     "keywords": [
-        "markdown"
+        "markdown-links"
     ],
     "author": "Tania Infante Bonilla",
     "license": "ISC",
@@ -31,11 +31,11 @@
         "cross-env": "^7.0.3",
         "figlet": "^1.5.2",
         "gradient-string": "^2.0.2",
-        "jest": "^29.2.1",
         "marked": "^4.1.1",
         "node-fetch": "^2.6.7"
     },
     "devDependencies": {
-        "ts-node": "^10.9.1"
+        "ts-node": "^10.9.1",
+        "jest": "^29.2.1"
         }
 }
