+++ test/index.spec.js
@@ -1,7 +1,11 @@
 const { mdLinks } = require('../index.js');
 
-describe('mdLinks', () => {
+describe('Md-links', () => {
     it('Mdlinks should be a function', () => {
+        expect(typeof mdLinks).toBe('function')
+    });
+
+    it('Md-links should return an object', () => {
         expect(typeof mdLinks).toBe('function')
     });
 });
